audio_codec_worker告警处理

+++ services/engine/codec/audio/audio_codec_worker.cpp
@@ -382,7 +382,7 @@ void AudioCodecWorker::Dispose()
 bool AudioCodecWorker::Begin()
 {
     AVCODEC_LOGD("Worker begin enter");
-    for (uint32_t i = 0; i < bufferCount; i++) {
+    for (int16_t i = 0; i < bufferCount; i++) {
         inBufAvaIndexQue_.push(i);
     }
     isRunning = true;
